Extract `src/compile/scale/domain`
