+++ test/compile/scale/domain.test.ts
@@ -1,0 +1,257 @@
+/* tslint:disable:quotemark */
+
+
+import {assert} from 'chai';
+import domain from '../../../src/compile/scale/domain';
+import {SOURCE, SUMMARY} from '../../../src/data';
+
+import {parseUnitModel} from '../../util';
+
+describe('compile/scale', () => {
+  describe('domain()', function() {
+    it('should return domain for stack', function() {
+      const model = parseUnitModel({
+        mark: "bar",
+        encoding: {
+          y: {
+            aggregate: 'sum',
+            field: 'origin',
+            type: 'quantitative'
+          },
+          x: {field: 'x', type: "ordinal"},
+          color: {field: 'color', type: "ordinal"},
+          row: {field: 'row', type: 'ordinal'}
+        }
+      });
+
+      const _domain = domain(model.scale('y'), model, 'y');
+
+      assert.deepEqual(_domain, {
+        data: 'stacked',
+        fields: ['sum_origin_start', 'sum_origin_end']
+      });
+    });
+
+    describe('for quantitative', function() {
+      it('should return the right domain for binned Q',
+        function() {
+          const model = parseUnitModel({
+            mark: "point",
+            encoding: {
+              y: {
+                bin: {maxbins: 15},
+                field: 'origin',
+                scale: {useRawDomain: true},
+                type: "quantitative"
+              }
+            }
+          });
+          const _domain = domain(model.scale('y'), model, 'y');
+
+          assert.deepEqual(_domain, {
+            data: SOURCE,
+            fields: [
+              'bin_origin_start',
+              'bin_origin_end'
+            ]
+          });
+        });
+
+      it('should return the raw domain if useRawDomain is true for non-bin, non-sum Q',
+        function() {
+          const model = parseUnitModel({
+            mark: "point",
+            encoding: {
+              y: {
+                aggregate: 'mean',
+                field: 'origin',
+                scale: {useRawDomain: true},
+                type: "quantitative"
+              }
+            }
+          });
+          const _domain = domain(model.scale('y'), model, 'y');
+
+          assert.deepEqual(_domain.data, SOURCE);
+        });
+
+      it('should return the aggregate domain for sum Q',
+        function() {
+          const model = parseUnitModel({
+            mark: "point",
+            encoding: {
+              y: {
+                aggregate: 'sum',
+                field: 'origin',
+                scale: {useRawDomain: true},
+                type: "quantitative"
+              }
+            }
+          });
+          const _domain = domain(model.scale('y'), model, 'y');
+
+          assert.deepEqual(_domain.data, SUMMARY);
+        });
+
+      it('should return the right custom domain', () => {
+        const model = parseUnitModel({
+          mark: "point",
+          encoding: {
+            y: {
+              field: 'horsepower',
+              type: "quantitative",
+              scale: {domain: [0,200]}
+            }
+          }
+        });
+        const _domain = domain(model.scale('y'), model, 'y');
+
+        assert.deepEqual(_domain, [0, 200]);
+      });
+
+      it('should return the aggregated domain if useRawDomain is false', function() {
+          const model = parseUnitModel({
+            mark: "point",
+            encoding: {
+              y: {
+                aggregate: 'min',
+                field: 'origin',
+                scale: {useRawDomain: false},
+                type: "quantitative"
+              }
+            }
+          });
+          const _domain = domain(model.scale('y'), model, 'y');
+
+          assert.deepEqual(_domain.data, SUMMARY);
+        });
+    });
+
+    describe('for time', function() {
+      it('should return the raw domain if useRawDomain is true for raw T',
+        function() {
+          const model = parseUnitModel({
+            mark: "point",
+            encoding: {
+              y: {
+                field: 'origin',
+                scale: {useRawDomain: true},
+                type: "temporal"
+              }
+            }
+          });
+          const _domain = domain(model.scale('y'), model, 'y');
+
+          assert.deepEqual(_domain.data, SOURCE);
+        });
+
+      it('should return the raw domain if useRawDomain is true for year T',
+        function() {
+          const model = parseUnitModel({
+            mark: "point",
+            encoding: {
+              y: {
+                field: 'origin',
+                scale: {useRawDomain: true},
+                type: "temporal",
+                timeUnit: 'year'
+              }
+            }
+          });
+          const _domain = domain(model.scale('y'), model, 'y');
+
+          assert.deepEqual(_domain.data, SOURCE);
+          assert.operator(_domain.field.indexOf('year'), '>', -1);
+        });
+
+      it('should return the correct domain for month T',
+        function() {
+          const model = parseUnitModel({
+            mark: "point",
+            encoding: {
+              y: {
+                field: 'origin',
+                type: "temporal",
+                timeUnit: 'month'
+              }
+            }
+          });
+          const _domain = domain(model.scale('y'), model, 'y');
+
+          assert.deepEqual(_domain, { data: 'source', field: 'month_origin', sort: {field: 'month_origin', op: 'min',} });
+        });
+
+        it('should return the correct domain for yearmonth T',
+          function() {
+            const model = parseUnitModel({
+              mark: "point",
+              encoding: {
+                y: {
+                  field: 'origin',
+                  type: "temporal",
+                  timeUnit: 'yearmonth'
+                }
+              }
+            });
+            const _domain = domain(model.scale('y'), model, 'y');
+
+            assert.deepEqual(_domain, {
+              data: 'source', field: 'yearmonth_origin',
+              sort: {field: 'yearmonth_origin', op: 'min'}
+            });
+          });
+
+      it('should return the right custom domain with DateTime objects', () => {
+        const model = parseUnitModel({
+          mark: "point",
+          encoding: {
+            y: {
+              field: 'year',
+              type: "temporal",
+              scale: {domain: [{year: 1970}, {year: 1980}]}
+            }
+          }
+        });
+        const _domain = domain(model.scale('y'), model, 'y');
+
+        assert.deepEqual(_domain, [
+          new Date(1970, 0, 1).getTime(),
+          new Date(1980, 0, 1).getTime()
+        ]);
+      });
+    });
+
+    describe('for ordinal', function() {
+      it('should return correct domain with the provided sort property', function() {
+        const sortDef = {op: 'min' as 'min', field:'Acceleration'};
+        const model = parseUnitModel({
+            mark: "point",
+            encoding: {
+              y: { field: 'origin', type: "ordinal", sort: sortDef}
+            }
+          });
+
+        assert.deepEqual(domain(model.scale('y'), model, 'y'), {
+            data: "source",
+            field: 'origin',
+            sort: sortDef
+          });
+      });
+
+      it('should return correct domain without sort if sort is not provided', function() {
+        const model = parseUnitModel({
+          mark: "point",
+          encoding: {
+            y: { field: 'origin', type: "ordinal"}
+          }
+        });
+
+        assert.deepEqual(domain(model.scale('y'), model, 'y'), {
+          data: "source",
+          field: 'origin',
+          sort: true
+        });
+      });
+    });
+  });
+});
